Exercise 4.21

+++ part4/controllers/blogs.js
@@ -28,8 +28,17 @@ blogsRouter.post('/', async (request, response) => {
 });
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id);
-  response.status(204).end();
+  const blog = await Blog.findById(request.params.id);
+  console.log(typeof blog.user);
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  console.log(decodedToken.id);
+
+  if (blog.user.toString() === decodedToken.id) {
+    await Blog.findByIdAndRemove(request.params.id);
+    response.status(204).end();
+  } else {
+    response.status(401).json({ error: 'Unauthorized' });
+  }
 });
 
 blogsRouter.put('/:id', async (request, response) => {
